Create component and container folders, create nav component
