+++ src/Components/Login.js
@@ -5,10 +5,14 @@ import { saveToken } from '../logic/storage';
 
 const Login = () => {
   const [redirect, setRedirect] = useState();
+
   const location = useLocation();
   const path = location.pathname.split('/')[2];
+
   const signUp = path !== 'login';
-  const submitValue = path === 'login' ? 'Login' : 'Sign Up';
+  const submitValue = signUp ? 'Sign Up' : 'Login';
+  const endpoint = signUp ? 'users' : 'login';
+  const validatePassword = signUp ? <input id="rPassword" type="password" placeholder="reapeat password" /> : <div />;
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -16,8 +20,6 @@ const Login = () => {
     const password = e.target.parentElement[1].value;
     const rPassword = e.target.parentElement[2].value;
     if (signUp && password !== rPassword) return;
-    let endpoint = 'login';
-    if (signUp) endpoint = 'users';
     const response = await login(name, password, endpoint);
     if (response.token) {
       saveToken(response.token);
@@ -26,8 +28,6 @@ const Login = () => {
       // display error message
     }
   };
-
-  const validatePassword = signUp ? <input id="rPassword" type="password" placeholder="reapeat password" /> : <div />;
 
   return (
     <>
