+++ src/Components/Nav.js
@@ -1,0 +1,8 @@
+const Nav = () => {
+  console.log('placeholder');
+  return (
+    <div>Track.it</div>
+  );
+};
+
+export default Nav;
