+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import App from './App';
+import App from './Components/App';
 import rootReducer from './reducers/index';
 import './assets/reset.css';
